+++ src/main/java/com/yunhalee/walkerholic/order/dto/OrderCartDTO.java
@@ -52,7 +52,7 @@ public class OrderCartDTO {
             this.stock = orderItem.getProduct().getStock();
             this.productId = orderItem.getProduct().getId();
             this.productName = orderItem.getProduct().getName();
-            this.productPrice = orderItem.getProduct().getPrice();
+            this.productPrice = orderItem.getProduct().getPrice().floatValue();
             this.productDescription = orderItem.getProduct().getDescription();
             this.productBrand = orderItem.getProduct().getBrand();
             this.productImageUrl = orderItem.getProduct().getMainImageUrl();
