+++ src/main/java/com/yunhalee/walkerholic/product/domain/Product.java
@@ -107,7 +107,7 @@ public class Product extends BaseTimeEntity {
         return productInfo.getStock();
     }
 
-    public Float getPrice(){
+    public BigDecimal getPrice(){
         return productInfo.getPrice();
     }
