+++ src/main/java/com/yunhalee/walkerholic/product/dto/SimpleProductResponse.java
@@ -19,7 +19,7 @@ public class SimpleProductResponse {
 
     private Integer stock;
 
-    private Float price;
+    private BigDecimal price;
 
     private BigDecimal average;
